requirements/requirements-dev.txt
aiofiles==0.8.0 ; python_full_version >= "3.8.1" and python_version < "4.0"
annotated-types==0.7.0 ; python_full_version >= "3.8.1" and python_version < "4.0"
any-api==0.1.0.11 ; python_full_version >= "3.8.1" and python_version < "4.0"
appdirs==1.4.4 ; python_full_version >= "3.8.1" and python_full_version < "4.0.0"
async-timeout==4.0.3 ; python_full_version >= "3.8.1" and python_full_version <= "3.11.2"
atomicwrites==1.4.1 ; python_full_version >= "3.8.1" and python_full_version < "4.0.0" and sys_platform == "win32"
attrs==23.2.0 ; python_full_version >= "3.8.1" and python_full_version < "4.0.0"
autoflake==1.4 ; python_full_version >= "3.8.1" and python_full_version < "4.0.0"
black==20.8b1 ; python_full_version >= "3.8.1" and python_full_version < "4.0.0"
certifi==2024.6.2 ; python_full_version >= "3.8.1" and python_full_version < "4.0.0"
cfgv==3.4.0 ; python_full_version >= "3.8.1" and python_full_version < "4.0.0"
charset-normalizer==2.0.12 ; python_full_version >= "3.8.1" and python_full_version < "4.0.0"
click==7.1.2 ; python_full_version >= "3.8.1" and python_full_version < "4.0.0"
colorama==0.4.6 ; python_full_version >= "3.8.1" and python_full_version < "4.0.0" and sys_platform == "win32"
coverage==7.2.0 ; python_full_version >= "3.8.1" and python_full_version < "4.0.0"
distlib==0.3.8 ; python_full_version >= "3.8.1" and python_full_version < "4.0.0"
filelock==3.15.4 ; python_full_version >= "3.8.1" and python_full_version < "4.0.0"
flask==2.0.0 ; python_full_version >= "3.8.1" and python_full_version < "4.0.0"
gprof2dot==2024.6.6 ; python_full_version >= "3.8.1" and python_full_version < "4.0.0"
h11==0.14.0 ; python_full_version >= "3.8.1" and python_full_version < "4.0.0"
httpcore==0.12.3 ; python_full_version >= "3.8.1" and python_full_version < "4.0.0"
httptools==0.6.1 ; python_full_version >= "3.8.1" and python_full_version < "4.0.0"
httpx==0.16.1 ; python_full_version >= "3.8.1" and python_full_version < "4.0.0"
identify==2.5.36 ; python_full_version >= "3.8.1" and python_full_version < "4.0.0"
idna==3.7 ; python_full_version >= "3.8.1" and python_full_version < "4.0.0"
importlib-metadata==4.11.4 ; python_full_version >= "3.8.1" and python_full_version < "4.0.0"
iniconfig==2.0.0 ; python_full_version >= "3.8.1" and python_full_version < "4.0.0"
isort==5.6.4 ; python_full_version >= "3.8.1" and python_version < "4.0"
itsdangerous==2.2.0 ; python_full_version >= "3.8.1" and python_full_version < "4.0.0"
jinja2==3.1.4 ; python_full_version >= "3.8.1" and python_full_version < "4.0.0"
markupsafe==2.1.5 ; python_full_version >= "3.8.1" and python_full_version < "4.0.0"
multidict==5.2.0 ; python_full_version >= "3.8.1" and python_full_version < "4.0.0"
mypy-extensions==1.0.0 ; python_full_version >= "3.8.1" and python_full_version < "4.0.0"
nodeenv==1.9.1 ; python_full_version >= "3.8.1" and python_full_version < "4.0.0"
packaging==24.1 ; python_full_version >= "3.8.1" and python_full_version < "4.0.0"
pathspec==0.12.1 ; python_full_version >= "3.8.1" and python_full_version < "4.0.0"
platformdirs==4.2.2 ; python_full_version >= "3.8.1" and python_full_version < "4.0.0"
pluggy==1.5.0 ; python_full_version >= "3.8.1" and python_full_version < "4.0.0"
pre-commit==3.5.0 ; python_full_version >= "3.8.1" and python_full_version < "4.0.0"
py-cpuinfo==9.0.0 ; python_full_version >= "3.8.1" and python_full_version < "4.0.0"
py==1.11.0 ; python_full_version >= "3.8.1" and python_full_version < "4.0.0"
pydantic-core==2.18.4 ; python_full_version >= "3.8.1" and python_version < "4.0"
pydantic==2.7.4 ; python_full_version >= "3.8.1" and python_version < "4.0"
pyflakes==3.2.0 ; python_full_version >= "3.8.1" and python_full_version < "4.0.0"
pytest-asyncio==0.14.0 ; python_full_version >= "3.8.1" and python_full_version < "4.0.0"
pytest-benchmark==4.0.0 ; python_full_version >= "3.8.1" and python_full_version < "4.0.0"
pytest-mock==3.5.1 ; python_full_version >= "3.8.1" and python_full_version < "4.0.0"
pytest-profiling==1.7.0 ; python_full_version >= "3.8.1" and python_full_version < "4.0.0"
pytest==6.2.5 ; python_full_version >= "3.8.1" and python_full_version < "4.0.0"
python-multipart==0.0.5 ; python_full_version >= "3.8.1" and python_full_version < "4.0.0"
pyyaml==6.0 ; python_full_version >= "3.8.1" and python_full_version < "4.0.0"
redis==4.6.0 ; python_full_version >= "3.8.1" and python_full_version < "4.0.0"
regex==2024.5.15 ; python_full_version >= "3.8.1" and python_full_version < "4.0.0"
requests==2.26.0 ; python_full_version >= "3.8.1" and python_full_version < "4.0.0"
rfc3986[idna2008]==1.5.0 ; python_full_version >= "3.8.1" and python_full_version < "4.0.0"
sanic-routing==23.12.0 ; python_full_version >= "3.8.1" and python_full_version < "4.0.0"
sanic-testing==0.3.0 ; python_full_version >= "3.8.1" and python_full_version < "4.0.0"
sanic==21.3.4 ; python_full_version >= "3.8.1" and python_full_version < "4.0.0"
six==1.16.0 ; python_full_version >= "3.8.1" and python_full_version < "4.0.0"
sniffio==1.3.1 ; python_full_version >= "3.8.1" and python_full_version < "4.0.0"
starlette==0.14.1 ; python_full_version >= "3.8.1" and python_full_version < "4.0.0"
toml==0.10.2 ; python_full_version >= "3.8.1" and python_full_version < "4.0.0"
tornado==6.1 ; python_full_version >= "3.8.1" and python_full_version < "4.0.0"
typed-ast==1.5.5 ; python_full_version >= "3.8.1" and python_full_version < "4.0.0"
typing-extensions==4.12.2 ; python_full_version >= "3.8.1" and python_version < "4.0"
ujson==5.10.0 ; sys_platform != "win32" and implementation_name == "cpython" and python_full_version >= "3.8.1" and python_full_version < "4.0.0"
urllib3==1.26.19 ; python_full_version >= "3.8.1" and python_full_version < "4.0.0"
uvicorn==0.13.4 ; python_full_version >= "3.8.1" and python_full_version < "4.0.0"
uvloop==0.19.0 ; sys_platform != "win32" and implementation_name == "cpython" and python_full_version >= "3.8.1" and python_full_version < "4.0.0"
virtualenv==20.26.3 ; python_full_version >= "3.8.1" and python_full_version < "4.0.0"
websockets==8.1 ; python_full_version >= "3.8.1" and python_full_version < "4.0.0"
werkzeug==2.0.0 ; python_full_version >= "3.8.1" and python_full_version < "4.0.0"
zipp==3.19.2 ; python_full_version >= "3.8.1" and python_full_version < "4.0.0"
